PoC/astra-poc-vc/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain-openai
langgraph
langchain-core
yfinance
copilotkit
qdrant-client
openai
